+++ libgearman/worker.cc
@@ -217,23 +217,23 @@ gearman_worker_options_t gearman_worker_options(const gearman_worker_st *worker)
   memset(&options, 0, sizeof(gearman_worker_options_t));
 
   if (worker->options.allocated)
-    options|= GEARMAN_WORKER_ALLOCATED;
+    options|= (int)GEARMAN_WORKER_ALLOCATED;
   if (worker->options.non_blocking)
-    options|= GEARMAN_WORKER_NON_BLOCKING;
+    options|= (int)GEARMAN_WORKER_NON_BLOCKING;
   if (worker->options.packet_init)
-    options|= GEARMAN_WORKER_PACKET_INIT;
+    options|= (int)GEARMAN_WORKER_PACKET_INIT;
   if (worker->options.grab_job_in_use)
-    options|= GEARMAN_WORKER_GRAB_JOB_IN_USE;
+    options|= (int)GEARMAN_WORKER_GRAB_JOB_IN_USE;
   if (worker->options.pre_sleep_in_use)
-    options|= GEARMAN_WORKER_PRE_SLEEP_IN_USE;
+    options|= (int)GEARMAN_WORKER_PRE_SLEEP_IN_USE;
   if (worker->options.work_job_in_use)
-    options|= GEARMAN_WORKER_WORK_JOB_IN_USE;
+    options|= (int)GEARMAN_WORKER_WORK_JOB_IN_USE;
   if (worker->options.change)
-    options|= GEARMAN_WORKER_CHANGE;
+    options|= (int)GEARMAN_WORKER_CHANGE;
   if (worker->options.grab_uniq)
-    options|= GEARMAN_WORKER_GRAB_UNIQ;
+    options|= (int)GEARMAN_WORKER_GRAB_UNIQ;
   if (worker->options.timeout_return)
-    options|= GEARMAN_WORKER_TIMEOUT_RETURN;
+    options|= (int)GEARMAN_WORKER_TIMEOUT_RETURN;
 
   return options;
 }
